requirements.txt
entsoe-py==0.6.12
prefect==2.19.0
pandas==2.2.0
pyarrow==15.0.0
python-dotenv==1.0.0
loguru==0.7.2